Shifted Profiler class to misc.profiler.py
